Added some basic input validation and HTML sanitization.
* HTML in the username (ughhhh....) will not be rendered on the page.
* Usernames must be at least three characters, and passwords 8.
* Both usernames and passwords can not be longer than 50 characters.
* Error messages for various pages are now shown on the page.

Passwords can not be longer than 50 characters because BCrypt is
guaranteed (with respect to various interpretations of the spec)
to usefully use at least 50 characters worth of information. Since the
rest may be thrown away, and 50 is large enough, this is a satisfactory
limit.

Error messages are currently abusing redirects and the session variable
to simply display error messages. This is understandably not ideal,
but can be revisited at another date.