Added rate limiting.
* Globally, 1000 requests can be made per hour per IP address.
* Login attempts don't allow more than 10 attempts per username per IP
before limiting kicks in.
* Added an automatic kill switch to the server if more than 25 games
get created.

The kill switch is a really simple measure to ensure too much CPU/memory
aren't eaten by the Node process, enabling me to login to the host
and recover it and do other things. This kill switch will definitely
be removed in the future, but is a nice protection against spammers
for now.